+++ src/app/core/type/app.type.ts
@@ -3,3 +3,9 @@ export enum VisibleFlag {
     Game = 1 << 0,
     Dialog = 1 << 1
 }
+
+export enum AppUpdateState {
+    None = 1,
+    Available = 2,
+    Downloaded = 3
+}