+++ src/app/layout/component/user-settings-dialog/user-settings-dialog.component.spec.ts
@@ -13,7 +13,10 @@ class ElectronProviderFake {
   }
 
   public provideIpcRenderer(): Electron.IpcRenderer {
-    return null;
+    return {
+      once: () => null,
+      send: () => null
+    } as any as Electron.IpcRenderer;
   }
 }