+++ src/app/modules/misc/service/misc-wiki.service.ts
@@ -1,11 +1,9 @@
 import { Injectable } from '@angular/core';
 import { BrowserService } from '@app/service';
-import { environment } from '@env/environment';
 import { SnackBarService } from '@shared/module/material/service';
-import { ItemClipboardResultCode, ItemClipboardService } from '@shared/module/poe/service';
-import { ItemService } from '@shared/module/poe/service/item/item.service';
-import { ItemSection, Language } from '@shared/module/poe/type';
-import { Observable, throwError, of } from 'rxjs';
+import { ItemClipboardResultCode, ItemClipboardService, ItemExternalService } from '@shared/module/poe/service';
+import { ItemSection } from '@shared/module/poe/type';
+import { Observable, of, throwError } from 'rxjs';
 import { catchError, flatMap } from 'rxjs/operators';
 
 @Injectable({
@@ -15,7 +13,7 @@ export class MiscWikiService {
 
     constructor(
         private readonly itemClipboard: ItemClipboardService,
-        private readonly itemService: ItemService,
+        private readonly itemExternalService: ItemExternalService,
         private readonly browser: BrowserService,
         private readonly snackbar: SnackBarService) {
     }
@@ -24,21 +22,18 @@ export class MiscWikiService {
         return this.itemClipboard.copy({
             [ItemSection.Rartiy]: true
         }).pipe(
-            flatMap(result => {
-                switch (result.code) {
+            flatMap(({ code, item }) => {
+                switch (code) {
                     case ItemClipboardResultCode.Success:
-                        let search = this.itemService.getName(result.item.nameId, Language.English) || '';
-                        if (search.length === 0) {
-                            search = this.itemService.getType(result.item.typeId, Language.English);
-                        }
-                        this.browser.open(`${environment.wiki.baseUrl}/index.php?search=${encodeURIComponent(search)}`, external);
+                        const url = this.itemExternalService.getWikiUrl(item);
+                        this.browser.open(url, external);
                         return of(null);
                     case ItemClipboardResultCode.Empty:
                         return this.snackbar.warning('clipboard.empty');
                     case ItemClipboardResultCode.ParserError:
                         return this.snackbar.warning('clipboard.parser-error');
                     default:
-                        return throwError(`code: '${result.code}' out of range`);
+                        return throwError(`code: '${code}' out of range`);
                 }
             }),
             catchError(() => {